Split 'HelperTrait' into generic and Drupal-specific helper traits.
